Normalization Lab 1
Did not understand what data was supposed to go in the two new columns
we created. Think I did ok in the early stages of the lab, was very
confused at the end though. Could not figure out what to use for primary
keys in the new tables. Will do better for next lab.